# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Delivery",
        "description": "API del Sistema de Gestión de Delivery",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/estado-pagos": {
            "get": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Listar todos los estados de pago",
                "operationId": "daba8934bb2d4988b628a9dc4be8c92f",
                "responses": {
                    "200": {
                        "description": "Estados de pagos obtenidos exitosamente."
                    },
                    "404": {
                        "description": "No se encontraron estados de pagos."
                    },
                    "500": {
                        "description": "Error interno del servidor."
                    }
                }
            },
            "post": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Crear un nuevo estado de pago",
                "operationId": "a0c6e2abf363363469ed9b1a6fb4d351",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Aprobado",
                                        "maxLength": 50
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "El pago fue procesado correctamente",
                                        "nullable": true,
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estado de pago creado exitosamente."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno al crear el estado de pago."
                    }
                }
            }
        },
        "/api/estado-pagos-desactivados": {
            "get": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Listar todos los estados de pago desactivados",
                "operationId": "267ed1939479910ce4b39c8476bfd8d3",
                "responses": {
                    "200": {
                        "description": "Estados de pagos desactivados obtenidos exitosamente."
                    },
                    "404": {
                        "description": "No se encontraron estados de pagos desactivados."
                    },
                    "500": {
                        "description": "Error interno del servidor al obtener estados de pago desactivados."
                    }
                }
            }
        },
        "/api/estado-pagos/{id}": {
            "get": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Obtener un estado de pago por ID",
                "operationId": "303fae5160d23fbc12e73ceb095cd6d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estado de pago",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de pago obtenido exitosamente."
                    },
                    "404": {
                        "description": "Estado de pago no encontrado."
                    },
                    "500": {
                        "description": "Error interno al obtener el estado de pago."
                    }
                }
            },
            "put": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Actualizar un estado de pago por ID",
                "operationId": "1cf290e524676817615a334514bd52fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estado de pago a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Rechazado",
                                        "maxLength": 50,
                                        "minLength": 4
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "El pago fue rechazado",
                                        "nullable": true,
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado de pago actualizado exitosamente."
                    },
                    "404": {
                        "description": "Estado de pago no encontrado."
                    },
                    "422": {
                        "description": "Error de validación."
                    },
                    "500": {
                        "description": "Error interno al actualizar el estado de pago."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Eliminar un estado de pago por ID",
                "operationId": "284db45dbf0db0f4c24867eb2597a151",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estado de pago a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de pago eliminado exitosamente."
                    },
                    "404": {
                        "description": "Estado de pago no encontrado."
                    },
                    "409": {
                        "description": "No se puede eliminar el estado de pago porque está en uso."
                    },
                    "422": {
                        "description": "ID no válido."
                    },
                    "500": {
                        "description": "Error interno al eliminar el estado de pago."
                    }
                }
            }
        },
        "/api/estado-pagos/{id}/restore": {
            "put": {
                "tags": [
                    "Estados de Pago"
                ],
                "summary": "Restaurar un estado de pago eliminado",
                "operationId": "cb41bf27a16de2ff2ee47473a90176ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del estado de pago a restaurar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de pago restaurado correctamente."
                    },
                    "404": {
                        "description": "Estado de pago no encontrado."
                    },
                    "409": {
                        "description": "El estado de pago no está eliminado."
                    },
                    "422": {
                        "description": "El ID no es válido."
                    },
                    "500": {
                        "description": "Error interno al restaurar el estado de pago."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación generado por Laravel Sanctum.",
                "bearerFormat": "Token Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}